Add C++ coordinate file output manager
Adds the ICoordinateOutput interface to organize coordinate output file
writing, as well as enum flags to communicate requirements between output
modules and output filetypes, and the outputmanager class to organize
the writing of new coordinate files.

OutputManager also allows the chaining of modules derived from
ICoordinateOutput together to perform actions that modify the meta
information contained in a t_trxframe before writing the information to
disk.

ICoordinateOutput can be used to communicate what requirements are specific
file writing method has on the meta information contained in t_trxframe,
with modules being able to proclaim those and outputmanager checking
them against the requested file types before accepting the addition of a
module to the processing chain.

+++ src/gromacs/fileio/tngio.cpp
@@ -1004,7 +1004,7 @@ void gmx_prepare_tng_writing(const char              *filename,
                              gmx_tng_trajectory_t    *gmx_tng_output,
                              int                      nAtoms,
                              const gmx_mtop_t        *mtop,
-                             const int               *index,
+                             const size_t            *index,
                              const char              *indexGroupName)
 {
 #if GMX_USE_TNG
@@ -1284,7 +1284,7 @@ real getDistanceScaleFactor(gmx_tng_trajectory_t in)
 
 void gmx_tng_setup_atom_subgroup(gmx_tng_trajectory_t gmx_tng,
                                  const int            nind,
-                                 const int           *ind,
+                                 const size_t        *ind,
                                  const char          *name)
 {
 #if GMX_USE_TNG
